pyproject.toml
[project]
name = "ml-infra"
version = "0.0.1"
description = "ML Infrastructure Project"
requires-python = ">=3.11"
dependencies = [
    "torch==2.8.0",
    "torchvision>=0.19.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "wandb>=0.21.0",
    "rich>=13.7.1",
    "requests>=2.32.0"
]

[project.scripts]
submit = "submit.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["submit", "ml"]
